SYNC: MMOCore r792 / KenM / 2007-05-06 15:13:40
MMOCore
This is still pending cleanup and opts for this very reason please do not send contribs to this now(they wont be accepted unless its a bugfix).
Notice that this project may go on a unstable state, so its encoraged to use the compiled jar that come with l2j itself witch is suposed to be the lastest stable version.